backend/requirements.txt
# BoxBrain backend — local-first, open source
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Vector search
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Media (frame extraction via ffmpeg CLI; optional helper)
ffmpeg-python>=0.2.0

# Vision: open-source image captioning (BLIP) for object/description from frames
transformers>=4.36.0
torch>=2.0.0
Pillow>=10.0.0

# Config
pydantic-settings>=2.1.0
